cambios realizados

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import { prisma } from '@/lib/prisma';
 import { NextResponse } from 'next/server'
 import * as yup from 'yup';
@@ -33,12 +34,18 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) { 
+
+    const user = await getUserSessionServer();
+
+    if(!user){
+        return NextResponse.json('No autorizado', {status: 401});
+    }
+
     try {
-        const body = await request.json();
 
-        const validatedData = await postSchema.validate(body);
+        const {complete, description} = await postSchema.validate(await request.json());
 
-        const todo = await prisma.todo.create({ data: validatedData });
+        const todo = await prisma.todo.create({ data: {complete, description, userId: user.id} });
 
         return NextResponse.json(todo);
     } catch (error) {
@@ -50,8 +57,14 @@ export async function POST(request: Request) {
 
 
 export async function DELETE() { 
+    const user = await getUserSessionServer();
+
+    if(!user){
+        return NextResponse.json('No autorizado', {status: 401});
+    }
+
     try {    
-      await prisma.todo.deleteMany({ where: { complete: true } }); 
+      await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } }); 
       return NextResponse.json({ message: 'Tareas completadas eliminadas' });
     } catch (error) {
       return NextResponse.json({ error }, { status: 400 });
